+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+**Breaking Changes**:
+
+- Remove the AWS lambda extension. ([#3568](https://github.com/getsentry/relay/pull/3568))
+
 **Bug fixes:**
 
 - Properly handle AI metrics from the Python SDK's `@ai_track` decorator ([#3539](https://github.com/getsentry/relay/pull/3539))
@@ -19,6 +23,8 @@
 - Forward `span.data` on the Kafka message. ([#3523](https://github.com/getsentry/relay/pull/3523))
 - Tag span duration metric like exclusive time. ([#3524](https://github.com/getsentry/relay/pull/3524))
 - Emit negative outcomes for denied metrics. ([#3508](https://github.com/getsentry/relay/pull/3508))
+- Increase size limits for internal batch endpoints. ([#3562](https://github.com/getsentry/relay/pull/3562))
+- Emit negative outcomes when metrics are rejected because of a disabled namespace. ([#3544](https://github.com/getsentry/relay/pull/3544))
 
 ## 24.4.2